requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0

numpy~=1.23.4
matplotlib~=3.6.1